+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Binary support in AWS Lambda/API Gateway example
-- Build and release tooling shell scripts and Dockerfile
+- Build and release tooling shell scripts and Dockerfile's
 - Integration tests and CircleCI setup
 - HeadlessChrome/62.0.3202.62 built for AWS Lambda 2017.03
 - serverless-plugin-chrome: support for limiting Chrome to only select service functions with the `custom.chrome.functions` parameter.